修改根据班课id查询视频信息列表

+++ src/main/java/com/teamsking/domain/entity/node/Node.java
@@ -225,7 +225,7 @@ public class Node {
      */
     private Integer deleteStatus;
 
-    private Double watchProgress;
+    private Integer watchProgress;
 
     private String suffixName;
 
@@ -819,11 +819,11 @@ public class Node {
         this.deleteStatus = deleteStatus;
     }
 
-    public Double getWatchProgress() {
+    public Integer getWatchProgress() {
         return watchProgress;
     }
 
-    public void setWatchProgress(Double watchProgress) {
+    public void setWatchProgress(Integer watchProgress) {
         this.watchProgress = watchProgress;
     }
